von int nach Long

+++ aero.minova.core.application.system.service/src/main/java/aero/minova/core/application/system/covid/test/print/domain/TestTermin.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TestTermin {
 	private Long keyLong;
-	private int CTSTeststreckeKey;
+	private long CTSTeststreckeKey;
 	private Long CTSTestpersonKey;
 	private String Starttime;
 }
